添加Zircon-VLIW处理器设计文档，包含架构概述、指令传递范式及相关模块的详细描述。同时，新增了相关的图形文件和更新了CPU、前端、后端等模块的实现，删除了不再使用的后端和转发模块。

+++ src/test/scala/simulator/Fetch.scala
@@ -17,6 +17,6 @@ class Fetch(mem: Memory, baseAddr: UInt) {
     }
 
     def fetch(): UInt = {
-        mem.read(pc, 2)
+        UInt(mem.read(pc, 2).toLong)
     }
 }
